pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openaustria-rag"
version = "0.1.0"
description = "Documentation platform with RAG-based gap analysis between code and documentation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "OpenAustria RAG Contributors"},
]

dependencies = [
    "llama-index>=0.11",
    "chromadb>=0.5",
    "tree-sitter>=0.21,<0.23",
    "tree-sitter-languages",
    "fastapi>=0.110",
    "uvicorn",
    "streamlit>=1.35",
    "gitpython>=3.1",
    "html2text",
    "requests",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.entry-points."openaustria_rag.connectors"]
git = "openaustria_rag.connectors.git_connector:GitConnector"
zip = "openaustria_rag.connectors.zip_connector:ZipConnector"
confluence = "openaustria_rag.connectors.confluence_connector:ConfluenceConnector"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: Integration tests spanning multiple components",
    "requires_ollama: Tests that require a running Ollama instance",
]

[tool.ruff]
target-version = "py311"
line-length = 100
